[feat/#71] pages폴더에 MainPage.jsx 파일 이동
